v055u

+++ app/src/full/java/com/celzero/bravedns/service/ProxyManager.kt
@@ -24,6 +24,7 @@ import com.celzero.bravedns.database.ProxyApplicationMapping
 import org.koin.core.component.KoinComponent
 import org.koin.core.component.inject
 import java.util.concurrent.CopyOnWriteArraySet
+import kotlin.collections.removeAll
 
 object ProxyManager : KoinComponent {
 
@@ -209,7 +210,12 @@ object ProxyManager : KoinComponent {
     }
 
     suspend fun updateApps(m: Collection<FirewallManager.AppInfoTuple>) {
-        m.forEach { updateApp(it.uid, it.packageName) }
+        m.forEach {
+            val newInfo = FirewallManager.getAppInfoByPackage(it.packageName) ?: return@forEach
+            if (newInfo.uid == it.uid) return@forEach // no change in uid
+
+            updateApp(newInfo.uid, it.packageName)
+        }
     }
 
     suspend fun addApp(appInfo: AppInfo?) {
@@ -282,10 +288,9 @@ object ProxyManager : KoinComponent {
     }
 
     private fun deleteFromCache(uid: Int, packageName: String) {
-        pamSet.forEach {
-            if (it.uid == uid && it.packageName == packageName) {
-                pamSet.remove(it)
-            }
+        val toRemove = pamSet.filter { it.uid == uid && it.packageName == packageName }
+        if (toRemove.isNotEmpty()) {
+            pamSet.removeAll(toRemove.toSet())
         }
     }
 
@@ -308,8 +313,12 @@ object ProxyManager : KoinComponent {
     }
 
     suspend fun deleteAppByPkgName(packageName: String) {
-        // delete the app from the cache
-        pamSet.removeIf { it.packageName == packageName }
+        val toRemove = pamSet.filter { it.packageName == packageName }
+        if (toRemove.isEmpty()) {
+            Logger.i(LOG_TAG_PROXY, "deleteAppByPkgName: app not found in proxy mapping: $packageName")
+            return
+        }
+        pamSet.removeAll(toRemove.toSet())
         // delete the app from the database
         db.deleteAppByPkgName(packageName)
         Logger.i(LOG_TAG_PROXY, "deleting app for mapping by package name: $packageName")
@@ -323,7 +332,11 @@ object ProxyManager : KoinComponent {
 
     suspend fun tombstoneApp(oldUid: Int) {
         // tombstone the app in the database and reload the cache
-        val newUid = -1 * oldUid // negative uid to indicate tombstone app
+        val newUid = if (oldUid > 0) -1 * oldUid else oldUid // negative uid to indicate tombstone app
+        if (newUid == oldUid) {
+            Logger.w(LOG_TAG_PROXY, "no change in uid, not tombstoning: $oldUid")
+            return
+        }
         db.tombstoneApp(oldUid, newUid)
         // reload the cache
         load()
